Modified code to implement task 0-3

+++ part3/hbnb/app/api/v1/reviews.py
@@ -3,6 +3,7 @@
 """
 from flask_restx import Namespace, Resource, fields
 from app.services.facade import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -14,15 +15,36 @@ review_model = api.model('Review', {
     'place_id': fields.String(required=True, description='ID of the place')
 })
 
+
 @api.route('/')
 class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(400, 'You cannot review your own place.')
+    @api.response(400, 'You have already reviewed this place.')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         # Placeholder for the logic to register a new review
+        current_user = get_jwt_identity()
         review_data = api.payload
+        review_data["user_id"] = current_user["id"]
+
+        if review_data['user_id'] != current_user["id"]:
+            return {'error': 'Unauthorized action'}, 403
+
+        user_places = facade.get_places_by_user_id(review_data["user_id"])
+        for place in user_places:
+            if review_data["place_id"] == place.id:
+                return {'error': 'You cannot review your own place.'}, 400 
+
+        place_reviews = facade.get_reviews_by_place(review_data["place_id"])
+        for review in place_reviews:
+            if review_data["user_id"] == review.user_id:
+                return {'error': 'You have already reviewed this place.'}, 400
+
 
         try:
             new_review = facade.create_review(review_data)
@@ -60,10 +82,17 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         # Placeholder for the logic to update a review by ID
+        current_user = get_jwt_identity()
         review_data = api.payload
+        review_data["user_id"] = current_user["id"]
+
+        if review_data['user_id'] != current_user["id"]:
+            return {'error': 'Unauthorized action'}, 403
 
         try:
             updated_review = facade.update_review(review_id, review_data)
@@ -76,10 +105,17 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
         # Placeholder for the logic to delete a review
+        current_user = get_jwt_identity()
         review = facade.get_review(review_id)
+
+        if review["user_id"] != current_user["id"]:
+            return {'error': 'Unauthorized action'}, 403
+
         if not review:
             return {'error': 'Review not found'}, 404
         facade.delete_review(review_id)
